Pad hex variants with '.' instead of '='
An important property of hex variants are they preserve bitwise
order, but if we use '=' (0x3d) it breaks this property.

For instance, with base64hex, depending on the input string length,
 bitwise sort of the result string could interchange inputs strings
ending (modulo 3 bytes) by 0xa0 or 0xa000:
* the string ending with 0xa0 would be transformed to   0x65303d3d
* the string ending with 0xa000 would be transformed to 0x6530303d
(NB the third byte difference) and consequently the second string will
be sorted before the first string if we compare the result strings, but
after the first string if we compare the input strings.

With a padding with '.', the two strings will be:
* 0x65302e2e
* 0x6530302e
And consequently the relative bitwise sort order is preserved.

Other choices than '.' are possible, the only constraint is it must be
before the character '0' in ASCII encoding (e.g. the space can be used).